+++ sentry-android-core/src/test/java/io/sentry/android/core/ActivityLifecycleIntegrationTest.kt
@@ -5,6 +5,7 @@ import android.app.Application
 import android.os.Bundle
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.check
+import com.nhaarman.mockitokotlin2.eq
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.never
 import com.nhaarman.mockitokotlin2.verify
@@ -17,10 +18,13 @@ import io.sentry.SentryOptions
 import io.sentry.SentryTracer
 import io.sentry.SpanStatus
 import io.sentry.TransactionContext
+import java.util.Date
+import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotNull
+import kotlin.test.assertSame
 import kotlin.test.assertTrue
 
 class ActivityLifecycleIntegrationTest {
@@ -29,19 +33,26 @@ class ActivityLifecycleIntegrationTest {
         val application = mock<Application>()
         val hub = mock<Hub>()
         val options = SentryAndroidOptions()
-        val activity = mock<Activity>()
         val bundle = mock<Bundle>()
-        val transaction = SentryTracer(TransactionContext("name", "op"), hub)
+        val context = TransactionContext("name", "op")
+        val transaction = SentryTracer(context, hub)
         val buildInfo = mock<IBuildInfoProvider>()
 
         fun getSut(apiVersion: Int = 29): ActivityLifecycleIntegration {
             whenever(hub.startTransaction(any<String>(), any())).thenReturn(transaction)
+            whenever(hub.options).thenReturn(options)
+            whenever(hub.startTransaction(any(), any(), any<Date>())).thenReturn(transaction)
             whenever(buildInfo.sdkInfoVersion).thenReturn(apiVersion)
             return ActivityLifecycleIntegration(application, buildInfo)
         }
     }
 
     private val fixture = Fixture()
+
+    @BeforeTest
+    fun `reset instance`() {
+        AppStartState.getInstance().resetInstance()
+    }
 
     @Test
     fun `When activity lifecycle breadcrumb is enabled, it registers callback`() {
@@ -100,6 +111,7 @@ class ActivityLifecycleIntegrationTest {
     fun `When ActivityBreadcrumbsIntegration is closed, it should unregister the callback`() {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
+
         sut.close()
 
         verify(fixture.application).unregisterActivityLifecycleCallbacks(any())
@@ -110,7 +122,9 @@ class ActivityLifecycleIntegrationTest {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityCreated(fixture.activity, fixture.bundle)
+        val activity = mock<Activity>()
+        sut.onActivityCreated(activity, fixture.bundle)
+
         verify(fixture.hub).addBreadcrumb(check<Breadcrumb> {
             assertEquals("ui.lifecycle", it.category)
             assertEquals("navigation", it.type)
@@ -123,7 +137,9 @@ class ActivityLifecycleIntegrationTest {
     fun `When activity is created, it should add a breadcrumb`() {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
-        sut.onActivityCreated(fixture.activity, fixture.bundle)
+
+        val activity = mock<Activity>()
+        sut.onActivityCreated(activity, fixture.bundle)
 
         verify(fixture.hub).addBreadcrumb(any<Breadcrumb>())
     }
@@ -132,7 +148,9 @@ class ActivityLifecycleIntegrationTest {
     fun `When activity is started, it should add a breadcrumb`() {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
-        sut.onActivityStarted(fixture.activity)
+
+        val activity = mock<Activity>()
+        sut.onActivityStarted(activity)
 
         verify(fixture.hub).addBreadcrumb(any<Breadcrumb>())
     }
@@ -142,7 +160,9 @@ class ActivityLifecycleIntegrationTest {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityResumed(fixture.activity)
+        val activity = mock<Activity>()
+        sut.onActivityResumed(activity)
+
         verify(fixture.hub).addBreadcrumb(any<Breadcrumb>())
     }
 
@@ -151,7 +171,9 @@ class ActivityLifecycleIntegrationTest {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPaused(fixture.activity)
+        val activity = mock<Activity>()
+        sut.onActivityPaused(activity)
+
         verify(fixture.hub).addBreadcrumb(any<Breadcrumb>())
     }
 
@@ -160,7 +182,9 @@ class ActivityLifecycleIntegrationTest {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityStopped(fixture.activity)
+        val activity = mock<Activity>()
+        sut.onActivityStopped(activity)
+
         verify(fixture.hub).addBreadcrumb(any<Breadcrumb>())
     }
 
@@ -169,7 +193,9 @@ class ActivityLifecycleIntegrationTest {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivitySaveInstanceState(fixture.activity, fixture.bundle)
+        val activity = mock<Activity>()
+        sut.onActivitySaveInstanceState(activity, fixture.bundle)
+
         verify(fixture.hub).addBreadcrumb(any<Breadcrumb>())
     }
 
@@ -178,7 +204,9 @@ class ActivityLifecycleIntegrationTest {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityDestroyed(fixture.activity)
+        val activity = mock<Activity>()
+        sut.onActivityDestroyed(activity)
+
         verify(fixture.hub).addBreadcrumb(any<Breadcrumb>())
     }
 
@@ -187,9 +215,11 @@ class ActivityLifecycleIntegrationTest {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
 
         verify(fixture.hub, never()).startTransaction(any<String>(), any())
+        verify(fixture.hub, never()).startTransaction(any(), any(), any<Date>())
     }
 
     @Test
@@ -198,24 +228,29 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+        sut.onActivityPreCreated(activity, fixture.bundle)
 
         // call only once
         verify(fixture.hub).startTransaction(any<String>(), any())
+        verify(fixture.hub, never()).startTransaction(any(), any(), any<Date>())
     }
 
     @Test
-    fun `Transaction op is navigation`() {
+    fun `Transaction op is ui_load`() {
         val sut = fixture.getSut()
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
+        setAppStartTime()
 
-        verify(fixture.hub).startTransaction(any<String>(), check {
-            assertEquals("navigation", it)
-        })
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+
+        verify(fixture.hub).startTransaction(any(), check {
+            assertEquals("ui.load", it)
+        }, any<Date>())
     }
 
     @Test
@@ -224,11 +259,14 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
+        setAppStartTime()
 
-        verify(fixture.hub).startTransaction(check<String> {
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+
+        verify(fixture.hub).startTransaction(check {
             assertEquals("Activity", it)
-        }, any())
+        }, any(), any<Date>())
     }
 
     @Test
@@ -246,7 +284,8 @@ class ActivityLifecycleIntegrationTest {
             assertNotNull(scope.transaction)
         }
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
     }
 
     @Test
@@ -259,14 +298,15 @@ class ActivityLifecycleIntegrationTest {
         whenever(fixture.hub.configureScope(any())).thenAnswer {
             val scope = Scope(fixture.options)
             val previousTransaction = SentryTracer(TransactionContext("name", "op"), fixture.hub)
-            scope.setTransaction(previousTransaction)
+            scope.transaction = previousTransaction
 
             sut.applyScope(scope, fixture.transaction)
 
             assertEquals(previousTransaction, scope.transaction)
         }
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
     }
 
     @Test
@@ -275,8 +315,9 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
-        sut.onActivityPostResumed(fixture.activity)
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+        sut.onActivityPostResumed(activity)
 
         verify(fixture.hub).captureTransaction(check {
             assertEquals(SpanStatus.OK, it.status)
@@ -289,11 +330,12 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
 
         fixture.transaction.status = SpanStatus.UNKNOWN_ERROR
 
-        sut.onActivityPostResumed(fixture.activity)
+        sut.onActivityPostResumed(activity)
 
         verify(fixture.hub).captureTransaction(check {
             assertEquals(SpanStatus.UNKNOWN_ERROR, it.status)
@@ -307,8 +349,9 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.isEnableActivityLifecycleTracingAutoFinish = false
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
-        sut.onActivityPostResumed(fixture.activity)
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+        sut.onActivityPostResumed(activity)
 
         verify(fixture.hub, never()).captureTransaction(any())
     }
@@ -318,7 +361,8 @@ class ActivityLifecycleIntegrationTest {
         val sut = fixture.getSut()
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPostResumed(fixture.activity)
+        val activity = mock<Activity>()
+        sut.onActivityPostResumed(activity)
 
         verify(fixture.hub, never()).captureTransaction(any())
     }
@@ -329,8 +373,9 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
-        sut.onActivityDestroyed(fixture.activity)
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+        sut.onActivityDestroyed(activity)
 
         verify(fixture.hub).captureTransaction(any())
     }
@@ -341,7 +386,8 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
 
         assertFalse(sut.activitiesWithOngoingTransactions.isEmpty())
     }
@@ -352,8 +398,9 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
-        sut.onActivityDestroyed(fixture.activity)
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+        sut.onActivityDestroyed(activity)
 
         assertTrue(sut.activitiesWithOngoingTransactions.isEmpty())
     }
@@ -364,10 +411,9 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        val activity = mock<Activity>()
-        sut.onActivityPreCreated(activity, mock())
+        sut.onActivityPreCreated(mock(), mock())
 
-        sut.onActivityPreCreated(fixture.activity, fixture.bundle)
+        sut.onActivityPreCreated(mock(), fixture.bundle)
         verify(fixture.hub).captureTransaction(any())
     }
 
@@ -381,6 +427,7 @@ class ActivityLifecycleIntegrationTest {
         sut.onActivityCreated(activity, mock())
 
         verify(fixture.hub, never()).startTransaction(any<String>(), any())
+        verify(fixture.hub, never()).startTransaction(any(), any(), any<Date>())
     }
 
     @Test
@@ -402,10 +449,12 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
+        setAppStartTime()
+
         val activity = mock<Activity>()
         sut.onActivityCreated(activity, mock())
 
-        verify(fixture.hub).startTransaction(any<String>(), any())
+        verify(fixture.hub).startTransaction(any(), any(), any<Date>())
     }
 
     @Test
@@ -419,5 +468,170 @@ class ActivityLifecycleIntegrationTest {
         sut.onActivityResumed(activity)
 
         verify(fixture.hub).captureTransaction(any())
+    }
+
+    @Test
+    fun `App start is Cold when savedInstanceState is null`() {
+        val sut = fixture.getSut(14)
+        fixture.options.tracesSampleRate = 1.0
+        sut.register(fixture.hub, fixture.options)
+
+        val activity = mock<Activity>()
+        sut.onActivityCreated(activity, null)
+
+        assertTrue(AppStartState.getInstance().isColdStart)
+    }
+
+    @Test
+    fun `App start is Warm when savedInstanceState is not null`() {
+        val sut = fixture.getSut(14)
+        fixture.options.tracesSampleRate = 1.0
+        sut.register(fixture.hub, fixture.options)
+
+        val activity = mock<Activity>()
+        val bundle = Bundle()
+        sut.onActivityCreated(activity, bundle)
+
+        assertFalse(AppStartState.getInstance().isColdStart)
+    }
+
+    @Test
+    fun `Do not overwrite App start type after set`() {
+        val sut = fixture.getSut(14)
+        fixture.options.tracesSampleRate = 1.0
+        sut.register(fixture.hub, fixture.options)
+
+        val activity = mock<Activity>()
+        val bundle = Bundle()
+        sut.onActivityCreated(activity, bundle)
+        sut.onActivityCreated(activity, null)
+
+        assertFalse(AppStartState.getInstance().isColdStart)
+    }
+
+    @Test
+    fun `App start end time is set`() {
+        val sut = fixture.getSut(14)
+        fixture.options.tracesSampleRate = 1.0
+        sut.register(fixture.hub, fixture.options)
+
+        setAppStartTime()
+
+        val activity = mock<Activity>()
+        sut.onActivityCreated(activity, null)
+        sut.onActivityResumed(activity)
+
+        // SystemClock.uptimeMillis() always returns 0, can't assert real values
+        assertNotNull(AppStartState.getInstance().appStartInterval)
+    }
+
+    @Test
+    fun `When firstActivityCreated is true, start transaction with given appStartTime`() {
+        val sut = fixture.getSut()
+        fixture.options.tracesSampleRate = 1.0
+        sut.register(fixture.hub, fixture.options)
+
+        val date = Date(0)
+        setAppStartTime(date)
+
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+
+        // call only once
+        verify(fixture.hub).startTransaction(any(), any(), eq(date))
+    }
+
+    @Test
+    fun `When firstActivityCreated is true, start app start warm span with given appStartTime`() {
+        val sut = fixture.getSut()
+        fixture.options.tracesSampleRate = 1.0
+        sut.register(fixture.hub, fixture.options)
+
+        val date = Date(0)
+        setAppStartTime(date)
+
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+
+        val span = fixture.transaction.children.first()
+        assertEquals(span.operation, "app.start.warm")
+        assertSame(span.startTimestamp, date)
+    }
+
+    @Test
+    fun `When firstActivityCreated is true, start app start cold span with given appStartTime`() {
+        val sut = fixture.getSut()
+        fixture.options.tracesSampleRate = 1.0
+        sut.register(fixture.hub, fixture.options)
+
+        val date = Date(0)
+        setAppStartTime(date)
+
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, null)
+
+        val span = fixture.transaction.children.first()
+        assertEquals(span.operation, "app.start.cold")
+        assertSame(span.startTimestamp, date)
+    }
+
+    @Test
+    fun `When firstActivityCreated is true, start app start span with Warm description`() {
+        val sut = fixture.getSut()
+        fixture.options.tracesSampleRate = 1.0
+        sut.register(fixture.hub, fixture.options)
+
+        val date = Date(0)
+        setAppStartTime(date)
+
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+
+        val span = fixture.transaction.children.first()
+        assertEquals(span.description, "Warm Start")
+        assertSame(span.startTimestamp, date)
+    }
+
+    @Test
+    fun `When firstActivityCreated is true, start app start span with Cold description`() {
+        val sut = fixture.getSut()
+        fixture.options.tracesSampleRate = 1.0
+        sut.register(fixture.hub, fixture.options)
+
+        val date = Date(0)
+        setAppStartTime(date)
+
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, null)
+
+        val span = fixture.transaction.children.first()
+        assertEquals(span.description, "Cold Start")
+        assertSame(span.startTimestamp, date)
+    }
+
+    @Test
+    fun `When firstActivityCreated is false, start transaction but not with given appStartTime`() {
+        val sut = fixture.getSut()
+        fixture.options.tracesSampleRate = 1.0
+        sut.register(fixture.hub, fixture.options)
+
+        setAppStartTime()
+
+        val activity = mock<Activity>()
+        sut.onActivityPreCreated(activity, fixture.bundle)
+
+        verify(fixture.hub).startTransaction(any(), any(), any<Date>())
+        sut.onActivityCreated(activity, fixture.bundle)
+        sut.onActivityPostResumed(activity)
+
+        val newActivity = mock<Activity>()
+        sut.onActivityPreCreated(newActivity, fixture.bundle)
+
+        verify(fixture.hub).startTransaction(any<String>(), any())
+    }
+
+    private fun setAppStartTime(date: Date = Date(0)) {
+        // set by SentryPerformanceProvider so forcing it here
+        AppStartState.getInstance().setAppStartTime(0, date)
     }
 }