+++ sentry/src/main/java/io/sentry/HubAdapter.java
@@ -3,6 +3,7 @@ package io.sentry;
 import io.sentry.protocol.SentryId;
 import io.sentry.protocol.SentryTransaction;
 import io.sentry.protocol.User;
+import java.util.Date;
 import java.util.List;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
@@ -172,6 +173,17 @@ public final class HubAdapter implements IHub {
       @Nullable CustomSamplingContext customSamplingContext,
       boolean bindToScope) {
     return Sentry.startTransaction(transactionContexts, customSamplingContext, bindToScope);
+  }
+
+  @ApiStatus.Internal
+  @Override
+  public @NotNull ITransaction startTransaction(
+      @NotNull TransactionContext transactionContexts,
+      @Nullable CustomSamplingContext customSamplingContext,
+      boolean bindToScope,
+      @Nullable Date startTimestamp) {
+    return Sentry.startTransaction(
+        transactionContexts, customSamplingContext, bindToScope, startTimestamp);
   }
 
   @Override