+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Feat: Measure app start time (#1487)
+
 ## 5.0.1
 
 * Fix: Sources and Javadoc artifacts were mixed up (#1515)